src/main/webapp/app/entities/user-mindmap/delete/user-mindmap-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { UserMindmapService } from '../service/user-mindmap.service';

import { UserMindmapDeleteDialogComponent } from './user-mindmap-delete-dialog.component';

describe('UserMindmap Management Delete Component', () => {
  let comp: UserMindmapDeleteDialogComponent;
  let fixture: ComponentFixture<UserMindmapDeleteDialogComponent>;
  let service: UserMindmapService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [UserMindmapDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(UserMindmapDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(UserMindmapDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(UserMindmapService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete('ABC');
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith('ABC');
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
